@@ -1,37 +1,47 @@
 {
-  "name": "react-router-5-example",
+  "name": "react-router-5",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "^24.0.11",
-    "@types/node": "^11.11.2",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router": "^5.1.13",
-    "@types/react-router-dom": "^5.3.3",
-    "query-string": "^6.12.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^5.3.3",
-    "react-router-dom": "^5.3.3",
-    "react-scripts": "^5.0.1",
-    "typescript": "^4.2.4",
-    "use-query-params": "*"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.11.45",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "5",
+    "react-router-dom": "5",
+    "react-scripts": "5.0.1",
+    "serialize-query-params": "file:.yalc/serialize-query-params",
+    "typescript": "^4.7.4",
+    "use-query-params": "file:.yalc/use-query-params",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "linkdeps": "yarn link ../../node_modules/react"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
