@@ -1,24 +1,31 @@
 {
-  "name": "react-router-6-example",
+  "name": "react-router-6",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "history": "^5.0.0",
-    "query-string": "^6.14.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.2",
-    "react-router": "^6.0.0-beta.0",
-    "react-router-dom": "^6.0.0-beta.0",
-    "react-scripts": "4.0.3",
-    "use-query-params": "*",
-    "web-vitals": "^1.0.1"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.11.45",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "query-string": "^7.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "serialize-query-params": "file:.yalc/serialize-query-params",
+    "typescript": "^4.7.4",
+    "use-query-params": "file:.yalc/use-query-params",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "linkdeps": "npm link ../../packages/use-query-params/node_modules/react"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -37,5 +44,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "resolutions": {
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.3.0",
+    "react-router": "6.3.0"
   }
 }
