@@ -1,15 +1,31 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import ReactDOM from 'react-dom/client';
+import './index.css';
 import App from './App';
 import { QueryParamProvider } from 'use-query-params';
-import './index.css';
+import { ReactRouterAdapter } from 'use-query-params/adapters/react-router';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+// optionally use the query-string parse / stringify functions to
+// handle more advanced cases than URLSearchParams supports.
+import { parse, stringify } from 'query-string';
 
-ReactDOM.render(
-  <Router>
-    <QueryParamProvider ReactRouterRoute={Route}>
-      <App />
-    </QueryParamProvider>
-  </Router>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+root.render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <QueryParamProvider
+        adapter={ReactRouterAdapter}
+        options={{
+          searchStringToObject: parse,
+          objectToSearchString: stringify,
+        }}
+      >
+        <Routes>
+          <Route path="/" element={<App />} />
+        </Routes>
+      </QueryParamProvider>
+    </BrowserRouter>
+  </React.StrictMode>
 );
