@@ -1,23 +1,27 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
-import App from './App';
 import { QueryParamProvider } from 'use-query-params';
 import { ReactRouter5Adapter } from 'use-query-params/adapters/react-router-5';
 import { parse, stringify } from 'query-string';
+import './index.css';
+import App from './App';
 
-ReactDOM.render(
-  <Router>
-    <QueryParamProvider
-      adapter={ReactRouter5Adapter}
-      options={{
-        enableBatching: true,
-        searchStringToObject: parse,
-        objectToSearchString: stringify,
-      }}
-    >
-      <App />
-    </QueryParamProvider>
-  </Router>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+root.render(
+  <React.StrictMode>
+    <Router>
+      <QueryParamProvider
+        adapter={ReactRouter5Adapter}
+        options={{
+          searchStringToObject: parse,
+          objectToSearchString: stringify,
+        }}
+      >
+        <App />
+      </QueryParamProvider>
+    </Router>
+  </React.StrictMode>
 );
